requirements.txt
openai>=1.0.0
pandas>=1.0.0
transformers>=4.0.0
torch>=1.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytest=>8.4.1
openpyxl>=3.0.0